+++ backend/src/main/java/com/backend/BackendApplication.java
@@ -3,9 +3,11 @@ package com.backend;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+import org.springframework.retry.annotation.EnableRetry;
 
 @SpringBootApplication
 @EnableJpaAuditing
+@EnableRetry
 public class BackendApplication {
 
     public static void main(String[] args) {
